[model] Allow indoor location without indoor coords #54
Fixed the issue by adding a confirmation script to the
loci.js which gives the opportunity to the admin to
confirm if he/she wish to save the indoor location
without indoor coords. Also modified the admin to make
sure this is saved without validation errors.
I also modified the model of object_location to allow
indoors to be None when type is indoor and no floorplan
is provided.

Fixes #54