[admin] Fixed ObjectLocation Admin for View only Permission #95
For view only permission, django admin uses 'AdminReadonlyField' for rendering data which looks for form fields in the model instance. And for widget to render 'read_only' must be set True. Setting the attributes to instance and 'read_only' geometry widget allowed the data to be viewed and map to render.

Fixes #95

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>