[changes] Removed floorplans on location change to outdoor #156
When user changes location from indoor to outdoor when floorplans were either existing or set in the form, it led to failure in saving form and floorplans related to Null location. Fixed this by controlling floorplans creation in location admin as no access to floorplan in location form. The same during location change is handled in respective forms by deleting related floorplans

Fixes #156

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>