[chore] Added mixin for handling fields for view only users. #95
To reduce duplicate code, added a mixin to handle fields which are required to be rendered explicitly for view-only users.
Followed consistent string quoting. Also renamed the test functions for better readability.
I also shifted the get_formset to base admin as it makes more sense to keep it part of base class

Fixes #95

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>