[fix] Fixed ObjectLocation admin read-only mode #95
In view-only mode, Django admin uses 'AdminReadonlyField' to render data.
This field checks the model instance for corresponding form fields.

To correctly render the geometry widget, its 'read_only' attribute must be
set to True. By setting the required attributes on the instance and marking
the geometry widget as read-only, the data is now visible and the map renders
properly.

Fixes #95

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>
Co-authored-by: Federico Capoano <f.capoano@openwisp.io>